Split Shiki highlighter from the plain token helpers
