+++ packages/gitbook/src/components/DocumentView/CodeBlock/plain-highlight.ts
@@ -10,7 +10,7 @@ import {
     getHighlightTokensText,
     parseDiffNotation,
     truncateHighlightTokens,
-} from './highlight';
+} from './highlight-tokens';
 
 /**
  * Parse a code block without highlighting it.
