+++ packages/gitbook/src/components/DocumentView/Prompt/Prompt.tsx
@@ -7,7 +7,7 @@ import {
 import { validateIconName } from '@gitbook/icons/icons';
 import type { BlockProps } from '../Block';
 import { CodeBlock } from '../CodeBlock';
-import { getPlainCodeBlock } from '../CodeBlock/highlight';
+import { getPlainCodeBlock } from '../CodeBlock/highlight-tokens';
 import { PromptClient } from './PromptClient';
 
 const PROMPT_COLLAPSED_LINE_COUNT = 10;
